requirements.txt
google-cloud-aiplatform>=1.36.0
langchain-core>=0.1.0
langchain-google-vertexai>=0.0.10
langchain-text-splitters>=0.0.1
langchain-community>=0.0.10
unstructured[all-docs]>=0.11.0
pypdf>=3.17.0
pydantic>=2.0.0
lxml>=4.9.0
pillow>=10.0.0
matplotlib>=3.7.0
opencv-python>=4.7.0
tiktoken>=0.4.0
python-dotenv>=1.0.0
